# backend/cmd/mainapp/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/v1/healthcheck": {
            "get": {
                "description": "현재 서버 상태 체크",
                "tags": [
                    "common"
                ],
                "summary": "상태 체크",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/v1/mc/status": {
            "get": {
                "description": "마인크래프트 서버 상태 확인",
                "tags": [
                    "mc"
                ],
                "summary": "서버 상태 확인",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/v1/mc/turn": {
            "post": {
                "description": "마인크래프트 서버 켜기/끄기",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mc"
                ],
                "summary": "서버 켜기/끄기",
                "parameters": [
                    {
                        "description": "서버 전원 ON/OFF",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.TurnMCRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        }
    },
    "definitions": {
        "models.TurnMCRequest": {
            "type": "object",
            "properties": {
                "turnOnOff": {
                    "type": "boolean"
                }
            }
        }
    }
}